backend/requirements.txt
# ── PyTorch (CPU-only, Python 3.13 compatible) ────────────────────────────────
# torch 2.6+ is the first release with officialPython 3.13 wheels.
# CPU wheels served from the extra index below ─ avoids 2 GB CUDA download.
# For GPU (CUDA 12.4): pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124
--extra-index-url https://download.pytorch.org/whl/cpu

# ── Core Framework ────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
asyncpg==0.30.0             # async driver → FastAPI / SQLAlchemy async engine
psycopg2-binary==2.9.10     # sync driver  → Celery tasks / Alembic migrations
alembic==1.14.0
motor==3.7.1                # async MongoDB driver
pymongo==4.11.1
# redis[asyncio] replaces both redis + the deprecated aioredis package
redis[asyncio]==5.2.1

# ── Task Queue ────────────────────────────────────────────────────────────────
celery==5.4.0
kombu==5.3.7

# ── Video & Audio Processing ──────────────────────────────────────────────────
yt-dlp>=2025.01.26
opencv-python-headless==4.10.0.84
moviepy==2.1.1
# faster-whisper replaces openai-whisper:
#   - Has proper Python 3.13 wheels (no C build required)
#   - openai-whisper on PyPI uses old setup.py broken on setuptools>=82
#   - whisper HEAD added numba dep which has no Python 3.13 wheels
#   - faster-whisper is 2-4x faster on CPU and has near-identical API
faster-whisper==1.1.1
# spleeter requires native C libs not available on Windows — install separately only if needed:
#   pip install spleeter
# spleeter==2.3.2

# ── Deep Learning (Python 3.13, CPU-only) ─────────────────────────────────────
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
transformers==4.48.0
tokenizers==0.21.0
accelerate==1.3.0
datasets==3.2.0
evaluate==0.4.3

# ── Computer Vision & OCR ─────────────────────────────────────────────────────
ultralytics==8.3.57     # YOLOv8 — 8.3.57 has Python 3.13 wheels
Pillow==10.4.0          # moviepy 2.x requires Pillow<11; 10.4.0 has Py3.13 wheels
pytesseract==0.3.13
scikit-image==0.25.1

# ── NLP ───────────────────────────────────────────────────────────────────────
spacy==3.8.11
en-core-web-sm==3.8.0
sentence-transformers==3.3.1
langdetect==1.0.9

# ── ML Utilities ──────────────────────────────────────────────────────────────
scikit-learn==1.6.1
numpy>=2.1.0
pandas==2.2.3
scipy==1.15.2

# ── External API Integrations ─────────────────────────────────────────────────
spotipy==2.24.0
httpx==0.28.1
aiohttp==3.11.11
requests==2.32.3

# ── Auth & Security ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-dotenv==1.0.1

# ── Observability ─────────────────────────────────────────────────────────────
loguru==0.7.3
prometheus-fastapi-instrumentator==7.0.0
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0

# ── File Handling & Export ────────────────────────────────────────────────────
reportlab==4.2.5        # PDF generation
openpyxl==3.1.5
# python-magic requires libmagic binary — use the Windows bundled version:
python-magic-bin==0.4.14; sys_platform == "win32"
python-magic==0.4.27; sys_platform != "win32"

# ── Caching & Serialization ───────────────────────────────────────────────────
orjson==3.10.12
cachetools==5.5.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
httpx==0.28.1
factory-boy==3.3.1